# [ConstraintSystem] Remove all stack allocation from `Solution`