@xedin 2023-08-30
@swift-ci please test compiler performance