@xedin 2023-08-31
# Summary for main full

**Regressions found (see below)**

- [Debug-batch](#debug-batch-53c8b0eca5869d40)
    - [debug-batch brief](#debug-batch-brief-53c8b0eca5869d40)
    - [debug-batch detailed](#debug-batch-detailed-53c8b0eca5869d40)
- [Release](#release-53c8b0eca5869d40)
    - [release brief](#release-brief-53c8b0eca5869d40)
    - [release detailed](#release-detailed-53c8b0eca5869d40)



<a name="debug-batch-53c8b0eca5869d40">

# Debug-batch

<a name="debug-batch-brief-53c8b0eca5869d40">

## debug-batch brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (2)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 137,924,561,274,041 | 138,169,559,250,468 | 244,997,976,427 | 0.18%
LLVM.NumLLVMBytesOutput | 2,475,092,882 | 2,482,419,608 | 7,326,726 | 0.3%
</details>

<a name="debug-batch-detailed-53c8b0eca5869d40">

## debug-batch detailed
<details>
<summary>Regressed (9)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ImportSetFoldHit | 384,431 | 388,394 | 3,962 | 1.03% :no_entry:
AST.ModuleVisibilityCacheMiss | 204,999 | 208,387 | 3,388 | 1.65% :no_entry:
AST.NumOperators | 838 | 850 | 11 | 1.37% :no_entry:
AST.NumPrecedenceGroups | 128 | 131 | 3 | 2.34% :no_entry:
Sema.GlobalActorInstanceRequest | 996 | 1,014 | 18 | 1.81% :no_entry:
Sema.LookupPrefixOperatorRequest | 346 | 351 | 5 | 1.45% :no_entry:
Sema.OpaqueResultTypeRequest | 504 | 510 | 6 | 1.19% :no_entry:
Sema.PropertyWrapperTypeInfoRequest | 580 | 594 | 13 | 2.33% :no_entry:
Sema.StructuralTypeRequest | 2,125 | 2,147 | 22 | 1.04% :no_entry:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (320)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ImportSetCacheHit | 3,932,224 | 3,942,519 | 10,295 | 0.26%
AST.ImportSetCacheMiss | 765,066 | 770,465 | 5,399 | 0.71%
AST.ImportSetFoldMiss | 380,634 | 382,071 | 1,437 | 0.38%
AST.ModuleShadowCacheHit | 66,504 | 65,945 | -559 | -0.84%
AST.ModuleShadowCacheMiss | 4,943 | 4,941 | -2 | -0.04%
AST.ModuleVisibilityCacheHit | 28,520,074 | 28,648,096 | 128,022 | 0.45%
AST.NumASTBytesAllocated | 59,860,129,652 | 59,939,829,105 | 79,699,453 | 0.13%
AST.NumASTScopeExpansions | 8,786,116 | 8,808,138 | 22,022 | 0.25%
AST.NumASTScopeLookups | 9,817,697 | 9,837,138 | 19,441 | 0.2%
AST.NumDecls | 343,215 | 344,215 | 1,000 | 0.29%
AST.NumDependencies | 386,752 | 387,711 | 959 | 0.25%
AST.NumIncrementalDependencies | 25,879 | 25,879 | 0 | 0.0%
AST.NumLinkLibraries | 0 | 0 | 0 | 0.0%
AST.NumLoadedModules | 312,956 | 313,900 | 943 | 0.3%
AST.NumLocalTypeDecls | 6,609 | 6,611 | 2 | 0.03%
AST.NumModuleLookupClassMember | 6,798 | 6,798 | 0 | 0.0%
AST.NumModuleLookupValue | 113,671,680 | 114,012,698 | 341,017 | 0.3%
AST.NumObjCMethods | 94,987 | 95,070 | 83 | 0.09%
AST.NumReferencedDynamicNames | 159 | 159 | 0 | 0.0%
AST.NumReferencedMemberNames | 15,098,062 | 15,144,190 | 46,128 | 0.31%
AST.NumReferencedTopLevelNames | 1,348,572 | 1,352,378 | 3,806 | 0.28%
AST.NumSourceBuffers | 395,711 | 396,853 | 1,142 | 0.29%
AST.NumSourceLines | 12,656,749 | 12,695,637 | 38,888 | 0.31%
AST.NumSourceLinesPerSecond | 4,701,028 | 4,707,294 | 6,265 | 0.13%
AST.NumTotalClangImportedEntities | 1,735,950 | 1,745,795 | 9,845 | 0.57%
Frontend.MaxMallocUsage | 0 | 0 | 0 | 0.0%
Frontend.NumInstructionsExecuted | 137,924,561,274,041 | 138,169,559,250,468 | 244,997,976,427 | 0.18%
Frontend.NumProcessFailures | 31 | 31 | 0 | 0.0%
IRGen.IRGenRequest | 28,745 | 28,829 | 84 | 0.29%
IRGen.OptimizedIRRequest | 0 | 0 | 0 | 0.0%
IRGen.SymbolObjectCodeRequest | 0 | 0 | 0 | 0.0%
IRModule.NumGOTEntries | 304,256 | 305,078 | 822 | 0.27%
IRModule.NumIRAliases | 247,855 | 248,679 | 823 | 0.33%
IRModule.NumIRBasicBlocks | 8,317,240 | 8,349,234 | 31,994 | 0.38%
IRModule.NumIRComdatSymbols | 0 | 0 | 0 | 0.0%
IRModule.NumIRFunctions | 4,556,869 | 4,569,596 | 12,727 | 0.28%
IRModule.NumIRGlobals | 4,286,847 | 4,302,923 | 16,076 | 0.38%
IRModule.NumIRIFuncs | 0 | 0 | 0 | 0.0%
IRModule.NumIRInsts | 88,148,793 | 88,446,688 | 297,894 | 0.34%
IRModule.NumIRNamedMetaData | 116,500 | 116,836 | 336 | 0.29%
IRModule.NumIRValueSymbols | 8,921,408 | 8,950,594 | 29,186 | 0.33%
LLVM.NumLLVMBytesOutput | 2,475,092,882 | 2,482,419,608 | 7,326,726 | 0.3%
Parse.ExportedSourceFileRequest | 4 | 4 | 0 | 0.0%
Parse.IDEInspectionSecondPassRequest | 0 | 0 | 0 | 0.0%
Parse.NumFunctionsParsed | 412,577 | 413,574 | 997 | 0.24%
Parse.NumIterableDeclContextParsed | 823,809 | 826,214 | 2,405 | 0.29%
Parse.ParseAbstractFunctionBodyRequest | 695,100 | 697,019 | 1,919 | 0.28%
Parse.ParseMembersRequest | 633,251 | 634,954 | 1,702 | 0.27%
Parse.ParseSourceFileRequest | 363,191 | 364,289 | 1,098 | 0.3%
Parse.ParseTopLevelDeclsRequest | 363,191 | 364,289 | 1,098 | 0.3%
SILGen.ASTLoweringRequest | 29,916 | 30,002 | 85 | 0.29%
SILGen.ParseSILModuleRequest | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenFunctions | 3,607,469 | 3,615,727 | 8,257 | 0.23%
SILModule.NumSILGenGlobalVariables | 126,539 | 127,700 | 1,160 | 0.92%
SILModule.NumSILGenVtables | 40,584 | 40,636 | 52 | 0.13%
SILModule.NumSILGenWitnessTables | 211,180 | 211,472 | 291 | 0.14%
SILModule.NumSILOptDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILOptFunctions | 4,086,959 | 4,097,155 | 10,195 | 0.25%
SILModule.NumSILOptGlobalVariables | 128,578 | 129,750 | 1,172 | 0.91%
SILModule.NumSILOptVtables | 40,953 | 41,005 | 52 | 0.13%
SILModule.NumSILOptWitnessTables | 254,924 | 255,344 | 420 | 0.16%
SILOptimizer.ExecuteSILPipelineRequest | 147,267 | 147,692 | 424 | 0.29%
SILOptimizer.LoweredSILRequest | 0 | 0 | 0 | 0.0%
Sema.ABIMembersRequest | 315,722 | 316,474 | 751 | 0.24%
Sema.AbstractGenericSignatureRequest | 148,805 | 148,773 | -32 | -0.02%
Sema.AccessLevelRequest | 18,919,751 | 18,864,612 | -55,138 | -0.29%
Sema.ActorIsolationRequest | 7,090,267 | 7,111,930 | 21,663 | 0.31%
Sema.AllMembersRequest | 176,781 | 177,233 | 452 | 0.26%
Sema.AnyObjectLookupRequest | 234 | 234 | 0 | 0.0%
Sema.ApplyAccessNoteRequest | 4,492,352 | 4,502,760 | 10,408 | 0.23%
Sema.AreAllStoredPropertiesDefaultInitableRequest | 50,092 | 50,137 | 45 | 0.09%
Sema.AttachedPropertyWrapperTypeRequest | 32,096 | 32,175 | 79 | 0.25%
Sema.AttachedPropertyWrappersRequest | 22,891,071 | 22,879,487 | -11,584 | -0.05%
Sema.AttachedResultBuilderRequest | 1,944,337 | 1,948,961 | 4,624 | 0.24%
Sema.BodyInitKindRequest | 98,675 | 98,814 | 139 | 0.14%
Sema.BreakTargetRequest | 6,826 | 6,839 | 13 | 0.2%
Sema.CXXNamespaceMemberLookup | 0 | 0 | 0 | 0.0%
Sema.CallerSideDefaultArgExprRequest | 495,980 | 496,127 | 146 | 0.03%
Sema.CheckDistributedFunctionRequest | 80 | 80 | 0 | 0.0%
Sema.CheckInconsistentAccessLevelOnImport | 58,014 | 58,184 | 170 | 0.29%
Sema.CheckInconsistentImplementationOnlyImportsRequest | 10,282 | 10,301 | 19 | 0.18%
Sema.CheckInconsistentSPIOnlyImportsRequest | 58,014 | 58,184 | 170 | 0.29%
Sema.CheckInconsistentWeakLinkedImportsRequest | 10,282 | 10,301 | 19 | 0.18%
Sema.CheckRedeclarationRequest | 2,393,442 | 2,399,974 | 6,532 | 0.27%
Sema.ClangCategoryLookupRequest | 0 | 0 | 0 | 0.0%
Sema.ClangDirectLookupRequest | 18,657 | 18,776 | 119 | 0.64%
Sema.ClangRecordMemberLookup | 18,657 | 18,776 | 119 | 0.64%
Sema.ClassAncestryFlagsRequest | 151,489 | 152,012 | 523 | 0.35%
Sema.ClosureEffectsRequest | 155,380 | 155,632 | 252 | 0.16%
Sema.ClosureHasExplicitResultRequest | 59,722 | 59,812 | 90 | 0.15%
Sema.CollectOverriddenDeclsRequest | 15,140,080 | 15,087,675 | -52,405 | -0.35%
Sema.CompareDeclSpecializationRequest | 875,259 | 875,767 | 508 | 0.06%
Sema.CompilerPluginLoadRequest | 4 | 4 | 0 | 0.0%
Sema.ConditionalRequirementsRequest | 890,604 | 891,733 | 1,129 | 0.13%
Sema.ConformanceHasEffectRequest | 63 | 63 | 0 | 0.0%
Sema.ConstantValueInfoRequest | 0 | 0 | 0 | 0.0%
Sema.ContinueTargetRequest | 2,277 | 2,280 | 3 | 0.13%
Sema.CursorInfoRequest | 0 | 0 | 0 | 0.0%
Sema.CustomAttrNominalRequest | 17,673 | 17,822 | 149 | 0.85%
Sema.CustomAttrTypeRequest | 3,480 | 3,492 | 12 | 0.34%
Sema.CustomRefCountingOperation | 0 | 0 | 0 | 0.0%
Sema.CxxRecordAsSwiftType | 0 | 0 | 0 | 0.0%
Sema.CxxRecordSemantics | 68,981 | 69,255 | 274 | 0.4%
Sema.DefaultAndMaxAccessLevelRequest | 122,900 | 123,260 | 360 | 0.29%
Sema.DefaultArgumentExprRequest | 109,321 | 109,833 | 511 | 0.47%
Sema.DefaultArgumentInitContextRequest | 859 | 859 | 0 | 0.0%
Sema.DefaultArgumentTypeRequest | 21,696 | 21,737 | 40 | 0.19%
Sema.DefaultDefinitionTypeRequest | 8,647 | 8,649 | 2 | 0.02%
Sema.DefaultTypeRequest | 670,467 | 671,620 | 1,153 | 0.17%
Sema.DerivativeAttrOriginalDeclRequest | 0 | 0 | 0 | 0.0%
Sema.DifferentiableAttributeTypeCheckRequest | 0 | 0 | 0 | 0.0%
Sema.DirectLookupRequest | 84,640,751 | 84,785,939 | 145,188 | 0.17%
Sema.DirectOperatorLookupRequest | 2,462,638 | 2,471,778 | 9,140 | 0.37%
Sema.DirectPrecedenceGroupLookupRequest | 923,541 | 926,707 | 3,166 | 0.34%
Sema.DistributedModuleIsAvailableRequest | 67 | 67 | 0 | 0.0%
Sema.DynamicallyReplacedDeclRequest | 7,563,859 | 7,586,543 | 22,684 | 0.3%
Sema.EnumElementExprPatternRequest | 19 | 19 | 0 | 0.0%
Sema.EnumRawTypeRequest | 42,576 | 42,724 | 147 | 0.35%
Sema.EnumRawValuesRequest | 24,880 | 24,964 | 84 | 0.34%
Sema.ExistentialConformsToSelfRequest | 33,964 | 34,025 | 61 | 0.18%
Sema.ExpandAccessorMacros | 1,079,705 | 1,082,895 | 3,189 | 0.3%
Sema.ExpandChildTypeRefinementContextsRequest | 710,113 | 712,544 | 2,430 | 0.34%
Sema.ExpandExtensionMacros | 1,039,030 | 1,038,867 | -163 | -0.02%
Sema.ExpandMacroExpansionDeclRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandMacroExpansionExprRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandMemberAttributeMacros | 68,657,158 | 68,604,703 | -52,455 | -0.08%
Sema.ExpandPeerMacroRequest | 49,149,876 | 49,026,602 | -123,273 | -0.25%
Sema.ExpandSynthesizedMemberMacroRequest | 317,269 | 318,088 | 819 | 0.26%
Sema.ExprPatternMatchRequest | 98,279 | 98,292 | 13 | 0.01%
Sema.ExtendedNominalRequest | 720,519 | 722,601 | 2,082 | 0.29%
Sema.ExtendedTypeRequest | 223,801 | 224,282 | 481 | 0.22%
Sema.ExternalMacroDefinitionRequest | 4 | 4 | 0 | 0.0%
Sema.FragileFunctionKindRequest | 2,850,008 | 2,857,657 | 7,649 | 0.27%
Sema.FunctionOperatorRequest | 19,695 | 19,740 | 45 | 0.23%
Sema.GenericParamListRequest | 7,282,112 | 7,295,431 | 13,318 | 0.18%
Sema.GenericSignatureRequest | 3,616,534 | 3,626,715 | 10,181 | 0.28%
Sema.GetDestructorRequest | 47,898 | 47,974 | 75 | 0.16%
Sema.GetDistributedActorArgumentDecodingMethodRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorIDPropertyRequest | 72 | 72 | 0 | 0.0%
Sema.GetDistributedActorImplicitCodableRequest | 117 | 117 | 0 | 0.0%
Sema.GetDistributedActorInvocationDecoderRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorSystemPropertyRequest | 63 | 63 | 0 | 0.0%
Sema.GetDistributedActorSystemRemoteCallFunctionRequest | 46 | 46 | 0 | 0.0%
Sema.GetDistributedRemoteCallArgumentInitFunctionRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedRemoteCallTargetInitFunctionRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedTargetInvocationDecoderDecodeNextArgumentFunctionRequest | 22 | 22 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordArgumentFunctionRequest | 20 | 20 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordErrorTypeFunctionRequest | 11 | 11 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordReturnTypeFunctionRequest | 16 | 16 | 0 | 0.0%
Sema.GetDistributedTargetInvocationResultHandlerOnReturnFunctionRequest | 12 | 12 | 0 | 0.0%
Sema.GetDistributedThunkRequest | 80 | 80 | 0 | 0.0%
Sema.GetImplicitSendableRequest | 255,199 | 255,593 | 394 | 0.15%
Sema.GetSourceFileAsyncNode | 138 | 138 | 0 | 0.0%
Sema.GlobalActorAttributeRequest | 9,010,327 | 9,037,204 | 26,877 | 0.3%
Sema.HasCircularInheritedProtocolsRequest | 18,882 | 18,927 | 45 | 0.24%
Sema.HasCircularRawValueRequest | 25,892 | 25,960 | 68 | 0.26%
Sema.HasDefaultInitRequest | 109,806 | 110,044 | 238 | 0.22%
Sema.HasDynamicCallableAttributeRequest | 31,890 | 32,032 | 142 | 0.45%
Sema.HasDynamicMemberLookupAttributeRequest | 577,079 | 578,661 | 1,581 | 0.27%
Sema.HasImportsMatchingFlagRequest | 720,021 | 722,217 | 2,196 | 0.3%
Sema.HasInitAccessorRequest | 5,549,866 | 5,545,960 | -3,906 | -0.07%
Sema.HasIsolatedSelfRequest | 16,007,699 | 16,020,259 | 12,560 | 0.08%
Sema.HasMemberwiseInitRequest | 54,981 | 55,127 | 146 | 0.27%
Sema.HasMissingDesignatedInitializersRequest | 33,153 | 33,234 | 81 | 0.24%
Sema.HasSelfOrAssociatedTypeRequirementsRequest | 16,846 | 16,904 | 57 | 0.34%
Sema.HasStorageRequest | 2,381,165 | 2,387,192 | 6,027 | 0.25%
Sema.HasUserDefinedDesignatedInitRequest | 109,816 | 110,054 | 238 | 0.22%
Sema.IDEInspectionFileRequest | 0 | 0 | 0 | 0.0%
Sema.ImplementsAttrProtocolRequest | 0 | 0 | 0 | 0.0%
Sema.InferredGenericSignatureRequest | 250,097 | 250,322 | 225 | 0.09%
Sema.InheritedDeclsReferencedRequest | 7,718,775 | 7,724,544 | 5,769 | 0.07%
Sema.InheritedProtocolsRequest | 834,914 | 836,763 | 1,849 | 0.22%
Sema.InheritedTypeRequest | 485,250 | 485,460 | 210 | 0.04%
Sema.InheritsSuperclassInitializersRequest | 41,816 | 41,924 | 107 | 0.26%
Sema.InitAccessorPropertiesRequest | 78,759 | 78,904 | 145 | 0.18%
Sema.InitAccessorReferencedVariablesRequest | 0 | 0 | 0 | 0.0%
Sema.InitKindRequest | 206,316 | 206,673 | 357 | 0.17%
Sema.InterfaceTypeRequest | 24,264,936 | 24,255,082 | -9,854 | -0.04%
Sema.IsABICompatibleOverrideRequest | 336,627 | 337,640 | 1,013 | 0.3%
Sema.IsAccessorTransparentRequest | 597,059 | 599,308 | 2,249 | 0.38%
Sema.IsActorRequest | 2,444,230 | 2,448,674 | 4,444 | 0.18%
Sema.IsCallAsFunctionNominalRequest | 6,854 | 6,872 | 18 | 0.27%
Sema.IsDeclApplicableRequest | 0 | 0 | 0 | 0.0%
Sema.IsDeclRefinementOfRequest | 69,872 | 70,131 | 259 | 0.37%
Sema.IsDefaultActorRequest | 64,197 | 64,317 | 120 | 0.19%
Sema.IsDistributedActorRequest | 2,729,528 | 2,733,916 | 4,388 | 0.16%
Sema.IsDynamicRequest | 3,390,103 | 3,399,076 | 8,973 | 0.26%
Sema.IsFinalRequest | 5,075,540 | 5,078,350 | 2,810 | 0.06%
Sema.IsGetterMutatingRequest | 848,896 | 851,723 | 2,827 | 0.33%
Sema.IsImplicitlyUnwrappedOptionalRequest | 4,163,510 | 4,173,818 | 10,308 | 0.25%
Sema.IsMoveOnlyRequest | 780,767 | 783,221 | 2,454 | 0.31%
Sema.IsNonUserModuleRequest | 103,116 | 103,101 | -15 | -0.02%
Sema.IsObjCRequest | 3,149,695 | 3,157,516 | 7,821 | 0.25%
Sema.IsSafeUseOfCxxDecl | 0 | 0 | 0 | 0.0%
Sema.IsSetterMutatingRequest | 761,742 | 764,208 | 2,466 | 0.32%
Sema.IsSingleValueStmtRequest | 392,543 | 393,213 | 670 | 0.17%
Sema.IsStaticRequest | 1,695,378 | 1,700,377 | 4,999 | 0.29%
Sema.LazyStoragePropertyRequest | 2,582 | 2,585 | 3 | 0.12%
Sema.LocalDiscriminatorsRequest | 579,312 | 580,309 | 997 | 0.17%
Sema.LookupAllConformancesInContextRequest | 3,120,747 | 3,112,188 | -8,559 | -0.27%
Sema.LookupConformanceInModuleRequest | 128,461,527 | 128,485,212 | 23,685 | 0.02%
Sema.LookupInModuleRequest | 9,815,039 | 9,836,263 | 21,224 | 0.22%
Sema.LookupInfixOperatorRequest | 81,054 | 81,317 | 263 | 0.33%
Sema.LookupPostfixOperatorRequest | 74 | 74 | 0 | 0.0%
Sema.LookupPrecedenceGroupRequest | 33,306 | 33,403 | 97 | 0.29%
Sema.MacroDefinitionRequest | 8 | 8 | 0 | 0.0%
Sema.MangleLocalTypeDeclRequest | 3,185 | 3,186 | 1 | 0.03%
Sema.MemberwiseInitPropertiesRequest | 78,761 | 78,906 | 145 | 0.18%
Sema.ModuleImplicitImportsRequest | 10,282 | 10,301 | 19 | 0.18%
Sema.ModuleLibraryLevelRequest | 22 | 22 | 0 | 0.0%
Sema.ModuleQualifiedLookupRequest | 3,665,929 | 3,672,181 | 6,252 | 0.17%
Sema.NamedLazyMemberLoadSuccessCount | 63,816,195 | 64,019,376 | 203,181 | 0.32%
Sema.NamingPatternRequest | 251,294 | 252,262 | 967 | 0.39%
Sema.NeedsNewVTableEntryRequest | 1,096,595 | 1,099,706 | 3,110 | 0.28%
Sema.NumAccessorBodiesSynthesized | 483,571 | 485,297 | 1,726 | 0.36%
Sema.NumAccessorsSynthesized | 584,256 | 586,454 | 2,198 | 0.38%
Sema.NumConformancePathsRecorded | 997,825 | 999,463 | 1,638 | 0.16%
Sema.NumConformancesDeserialized | 6,112,153 | 6,112,248 | 95 | 0.0%
Sema.NumConstraintScopes | 78,616,476 | 78,678,121 | 61,645 | 0.08%
Sema.NumConstraintsConsideredForEdgeContraction | 2,810,851 | 2,811,381 | 529 | 0.02%
Sema.NumCrossImportsChecked | 0 | 0 | 0 | 0.0%
Sema.NumCrossImportsFound | 0 | 0 | 0 | 0.0%
Sema.NumCyclicOneWayComponentsCollapsed | 0 | 0 | 0 | 0.0%
Sema.NumDeclsDeserialized | 95,855,275 | 95,834,010 | -21,265 | -0.02%
Sema.NumDeclsTypechecked | 3,395,250 | 3,405,106 | 9,855 | 0.29%
Sema.NumLazyIterableDeclContexts | 14,372,609 | 14,396,361 | 23,752 | 0.17%
Sema.NumLazyRequirementSignatures | 1,098,111 | 1,100,265 | 2,154 | 0.2%
Sema.NumLazyRequirementSignaturesLoaded | 773,155 | 774,681 | 1,526 | 0.2%
Sema.NumLeafScopes | 60,579,275 | 60,625,817 | 46,542 | 0.08%
Sema.NumRequirementMachineCompletionSteps | 26,867,469 | 26,886,819 | 19,350 | 0.07%
Sema.NumRequirementMachineUnifiedConcreteTerms | 1,333,913 | 1,332,877 | -1,036 | -0.08%
Sema.NumRequirementMachines | 2,290,430 | 2,291,851 | 1,421 | 0.06%
Sema.NumTypesDeserialized | 26,484,389 | 26,503,163 | 18,774 | 0.07%
Sema.NumUnloadedLazyIterableDeclContexts | 10,948,734 | 10,982,744 | 34,009 | 0.31%
Sema.ObjCInterfaceAndImplementationRequest | 955,207 | 959,840 | 4,633 | 0.49%
Sema.OpaqueReadOwnershipRequest | 515,140 | 517,203 | 2,063 | 0.4%
Sema.OperatorPrecedenceGroupRequest | 915 | 924 | 9 | 0.98%
Sema.OverriddenDeclsRequest | 6,664,716 | 6,672,748 | 8,032 | 0.12%
Sema.ParamSpecifierRequest | 2,309,210 | 2,315,725 | 6,515 | 0.28%
Sema.PatternBindingEntryRequest | 838,979 | 841,905 | 2,926 | 0.35%
Sema.PatternTypeRequest | 994,082 | 997,348 | 3,266 | 0.33%
Sema.PolymorphicEffectKindRequest | 1,135,176 | 1,137,945 | 2,769 | 0.24%
Sema.PolymorphicEffectRequirementsRequest | 75 | 75 | 0 | 0.0%
Sema.PotentialMacroExpansionsInContextRequest | 7,289,168 | 7,311,653 | 22,484 | 0.31%
Sema.PreCheckFunctionBodyRequest | 458,700 | 459,600 | 900 | 0.2%
Sema.PreCheckResultBuilderRequest | 954 | 959 | 5 | 0.52%
Sema.PreCheckReturnStmtRequest | 624,370 | 625,604 | 1,234 | 0.2%
Sema.PrimaryAssociatedTypesRequest | 19,586 | 19,554 | -32 | -0.17%
Sema.PrimarySourceFilesRequest | 10,282 | 10,301 | 19 | 0.18%
Sema.PropertyWrapperAuxiliaryVariablesRequest | 2,541,975 | 2,547,290 | 5,314 | 0.21%
Sema.PropertyWrapperBackingPropertyTypeRequest | 32,096 | 32,175 | 79 | 0.25%
Sema.PropertyWrapperInitializerInfoRequest | 734,811 | 737,136 | 2,325 | 0.32%
Sema.PropertyWrapperLValuenessRequest | 2,624 | 2,632 | 8 | 0.3%
Sema.PropertyWrapperMutabilityRequest | 842,017 | 844,834 | 2,817 | 0.33%
Sema.ProtocolDependenciesRequest | 662,672 | 663,901 | 1,229 | 0.19%
Sema.ProtocolRequirementsRequest | 82,697 | 82,872 | 174 | 0.21%
Sema.ProtocolRequiresClassRequest | 48,201 | 48,260 | 59 | 0.12%
Sema.ProvideDefaultImplForRequest | 15,140,080 | 15,087,675 | -52,405 | -0.35%
Sema.QualifiedLookupRequest | 10,290,295 | 10,308,461 | 18,165 | 0.18%
Sema.RangeInfoRequest | 0 | 0 | 0 | 0.0%
Sema.RawCommentRequest | 1,243,515 | 1,246,319 | 2,804 | 0.23%
Sema.RenamedDeclRequest | 5,545 | 5,545 | 0 | 0.0%
Sema.RequirementRequest | 196,470 | 196,661 | 191 | 0.1%
Sema.RequirementSignatureRequest | 832,649 | 834,265 | 1,616 | 0.19%
Sema.RequiresOpaqueAccessorsRequest | 2,159,880 | 2,165,412 | 5,532 | 0.26%
Sema.RequiresOpaqueModifyCoroutineRequest | 519,050 | 520,910 | 1,860 | 0.36%
…